Split shell process detection
